Remove deprecated API + queryState logic
