+++ packages/toolpad-app/pages/api/rpc.ts
@@ -5,7 +5,6 @@ import {
   getApps,
   getApp,
   createApp,
-  execApi,
   execQuery,
   dataSourceFetchPrivate,
   loadDom,
@@ -116,9 +115,6 @@ const rpcServer = {
     }),
     getApp: createMethod<typeof getApp>((params) => {
       return getApp(...params);
-    }),
-    execApi: createMethod<typeof execApi>((args) => {
-      return execApi(...args);
     }),
     execQuery: createMethod<typeof execQuery>((args) => {
       return execQuery(...args);
