+++ packages/toolpad-app/src/appDom.ts
@@ -37,13 +37,11 @@ export function compareFractionalIndex(index1: string, index2: string): number {
 type AppDomNodeType =
   | 'app'
   | 'connection'
-  | 'api'
   | 'theme'
   | 'page'
   | 'element'
   | 'codeComponent'
-  | 'query'
-  | 'queryState';
+  | 'query';
 
 interface AppDomNodeBase {
   readonly id: NodeId;
@@ -74,17 +72,6 @@ export interface ConnectionNode<P = unknown> extends AppDomNodeBase {
   };
 }
 
-export interface ApiNode<Q = unknown> extends AppDomNodeBase {
-  readonly type: 'api';
-  readonly attributes: {
-    readonly connectionId: ConstantAttrValue<string>;
-    readonly dataSource: ConstantAttrValue<string>;
-    readonly query: ConstantAttrValue<Q>;
-    readonly transform?: ConstantAttrValue<string>;
-    readonly transformEnabled?: ConstantAttrValue<boolean>;
-  };
-}
-
 export interface PageNode extends AppDomNodeBase {
   readonly type: 'page';
   readonly attributes: {
@@ -108,17 +95,6 @@ export interface CodeComponentNode extends AppDomNodeBase {
   };
 }
 
-export interface QueryStateNode<P = any> extends AppDomNodeBase {
-  readonly type: 'queryState';
-  readonly attributes: {
-    readonly api: ConstantAttrValue<NodeId | null>;
-    readonly refetchOnWindowFocus?: ConstantAttrValue<boolean>;
-    readonly refetchOnReconnect?: ConstantAttrValue<boolean>;
-    readonly refetchInterval?: ConstantAttrValue<number>;
-  };
-  readonly params?: BindableAttrValues<P>;
-}
-
 export interface QueryNode<Q = any, P = any> extends AppDomNodeBase {
   readonly type: 'query';
   readonly params?: BindableAttrValues<P>;
@@ -137,12 +113,10 @@ export interface QueryNode<Q = any, P = any> extends AppDomNodeBase {
 type AppDomNodeOfType<K extends AppDomNodeType> = {
   app: AppNode;
   connection: ConnectionNode;
-  api: ApiNode;
   theme: ThemeNode;
   page: PageNode;
   element: ElementNode;
   codeComponent: CodeComponentNode;
-  queryState: QueryStateNode;
   query: QueryNode;
 }[K];
 
@@ -150,23 +124,19 @@ type AllowedChildren = {
   app: {
     pages: 'page';
     connections: 'connection';
-    apis: 'api';
     themes: 'theme';
     codeComponents: 'codeComponent';
   };
   theme: {};
-  api: {};
   connection: {};
   page: {
     children: 'element';
-    queryStates: 'queryState';
     queries: 'query';
   };
   element: {
     [prop: string]: 'element';
   };
   codeComponent: {};
-  queryState: {};
   query: {};
 };
 
@@ -288,14 +258,6 @@ export function assertIsPage(node: AppDomNode): asserts node is PageNode {
   assertIsType<PageNode>(node, 'page');
 }
 
-export function isApi<P>(node: AppDomNode): node is ApiNode<P> {
-  return isType<ApiNode>(node, 'api');
-}
-
-export function assertIsApi<P>(node: AppDomNode): asserts node is ApiNode<P> {
-  assertIsType<ApiNode>(node, 'api');
-}
-
 export function isConnection<P>(node: AppDomNode): node is ConnectionNode<P> {
   return isType<ConnectionNode>(node, 'connection');
 }
@@ -326,14 +288,6 @@ export function isElement<P>(node: AppDomNode): node is ElementNode<P> {
 
 export function assertIsElement<P>(node: AppDomNode): asserts node is ElementNode<P> {
   assertIsType<ElementNode>(node, 'element');
-}
-
-export function isQueryState<P>(node: AppDomNode): node is QueryStateNode<P> {
-  return isType<QueryStateNode>(node, 'queryState');
-}
-
-export function assertIsQueryState<P>(node: AppDomNode): asserts node is QueryStateNode<P> {
-  assertIsType<QueryStateNode>(node, 'queryState');
 }
 
 export function isQuery<P>(node: AppDomNode): node is QueryNode<P> {
@@ -753,15 +707,7 @@ export function getNodeIdByName(dom: AppDom, name: string): NodeId | null {
  * TODO: Would it make sense to create a separate datastructure that represents the render tree?
  */
 export function createRenderTree(dom: AppDom): AppDom {
-  const frontendNodes = new Set([
-    'app',
-    'page',
-    'element',
-    'queryState',
-    'query',
-    'theme',
-    'codeComponent',
-  ]);
+  const frontendNodes = new Set(['app', 'page', 'element', 'query', 'theme', 'codeComponent']);
   return {
     ...dom,
     nodes: filterValues(dom.nodes, (node) => frontendNodes.has(node.type)) as AppDomNodes,
