+++ packages/toolpad-app/src/server/handleDataRequest.ts
@@ -1,6 +1,6 @@
 import { NextApiRequest, NextApiResponse } from 'next';
 import Cors from 'cors';
-import { execApi, execQuery, loadVersionedDom } from './data';
+import { execQuery, loadVersionedDom } from './data';
 import initMiddleware from './initMiddleware';
 import { NodeId, ApiResult, VersionOrPreview } from '../types';
 import * as appDom from '../appDom';
@@ -27,24 +27,13 @@ export default async (
   await cors(req, res);
   const queryNodeId = req.query.queryId as NodeId;
   const dom = await loadVersionedDom(appId, version);
-  const query = appDom.getNode(dom, queryNodeId);
+  const query = appDom.getNode(dom, queryNodeId, 'query');
 
-  let result;
-  if (appDom.isApi(query)) {
-    result = await execApi(
-      appId,
-      query,
-      req.query.params ? JSON.parse(req.query.params as string) : {},
-    );
-  } else if (appDom.isQuery(query)) {
-    result = await execQuery(
-      appId,
-      query,
-      req.query.params ? JSON.parse(req.query.params as string) : {},
-    );
-  } else {
-    throw new Error(`Unrecognized query "${queryNodeId}"`);
-  }
+  const result = await execQuery(
+    appId,
+    query,
+    req.query.params ? JSON.parse(req.query.params as string) : {},
+  );
 
   res.json(result);
 };
