+++ packages/toolpad-app/src/toolpadDataSources/client.tsx
@@ -1,9 +1,7 @@
-import { ArgTypeDefinitions } from '@mui/toolpad-core';
 import movies from './movies/client';
 import postgres from './postgres/client';
 import rest from './rest/client';
-import { NodeId, ClientDataSource } from '../types';
-import * as appDom from '../appDom';
+import { ClientDataSource } from '../types';
 import googleSheets from './googleSheets/client';
 
 const clientDataSources: { [key: string]: ClientDataSource<any, any> | undefined } = {
@@ -12,24 +10,5 @@ const clientDataSources: { [key: string]: ClientDataSource<any, any> | undefined
   rest,
   googleSheets,
 };
-
-export function getQueryNodeArgTypes(
-  dom: appDom.AppDom,
-  node: appDom.QueryStateNode,
-): ArgTypeDefinitions {
-  const apiNodeId = node.attributes.api.value;
-  if (!apiNodeId) {
-    return {};
-  }
-  const apiNode = apiNodeId ? appDom.getNode(dom, apiNodeId, 'api') : null;
-  if (!apiNode) {
-    console.warn(`Can't resolve API node "${apiNodeId}" from query "${node.id}"`);
-    return {};
-  }
-  const connectionNodeId = apiNode.attributes.connectionId.value as NodeId;
-  const connectionNode = appDom.getNode(dom, connectionNodeId, 'connection');
-  const dataSource = clientDataSources[connectionNode.attributes.dataSource.value];
-  return dataSource?.getArgTypes?.(apiNode.attributes.query.value) || {};
-}
 
 export default clientDataSources;
