+++ packages/toolpad-app/src/toolpadDataSources/rest/types.ts
@@ -31,10 +31,6 @@ export interface RestConnectionParams {
 }
 
 export interface FetchQuery {
-  /**
-   * @deprecated To be removed when QueryStateNode is removed
-   */
-  readonly params: Record<string, string>;
   readonly url: BindableAttrValue<string>;
   readonly method: string;
   readonly headers: [string, BindableAttrValue<string>][];
